Abstracted Join to Custom Component
